requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
redis
celery
google-generativeai
uvicorn
python-dotenv
mysqlclient